+++ lib/src/builders/class_builder.dart
